requirements-dev.txt
-e .
ckuehl-pre-commit-types==0.7.6.dev1
coverage
coveralls
ipdb
mock
pytest
pytest-django
